==> Natural_Language_Processing/Advanced_NLP_with_Python/01_Reading_text_data.ipynb <==
## NLP Basics: Reading in text data

#### Read In Semi-structured Text Data

import pandas as pd

messages = pd.read_csv('data/spam.csv', encoding='latin-1')
messages.head()

# Drop unused columns and label columns that will be used
messages = messages.drop(labels=["Unnamed: 2", "Unnamed: 3", "Unnamed: 4"], axis=1)
messages.columns = ['label', 'text']

messages.head()

messages.shape

messages.label.value_counts()

print(f"Number of nulls in label: {messages['label'].isnull().sum()}")
print(f"Number of nulls in text: {messages['text'].isnull().sum()}")
