==> Natural_Language_Processing/Advanced_NLP_with_Python/02_Cleaning_text_data.ipynb <==
## NLP Basics: Implementing A Pipeline To Clean Text

#### Preprocessing Text Data
Cleaning up the text data is necessary to highlight attributes that you're going to want your machine learning system to pick up on. We will explore three pre-processing steps in this notebook:
* Remove punctuation
* Tokenization
* Remove stopwords

import pandas as pd
import string
from nltk.corpus import stopwords
import re

messages = pd.read_csv('data/spam.csv', encoding='latin-1')
messages.head()

# Drop unused columns and label columns that will be used
messages = messages.drop(labels=["Unnamed: 2", "Unnamed: 3", "Unnamed: 4"], axis=1)
messages.columns = ['label', 'text']

messages.head()

### Remove Punctuation

def remove_punct(text):
    text = "".join([char for char in text if char not in string.punctuation])
    return text

messages['text_clean'] = messages['text'].apply(lambda x: remove_punct(x))

messages.head()

### Tokenize

def tokenize(text):
    return re.split('\W+', text)

messages['text_tokenized'] = messages['text_clean'].apply(lambda x: tokenize(x.lower()))

messages.head()

### Remove stopwords

stopword = stopwords.words('english')

def remove_stopwords(tokens):
    return [word for word in tokens if word not in stopword]

messages['text_nostop'] = messages['text_tokenized'].apply(lambda x: remove_stopwords(x))

messages.head()
