==> Natural_Language_Processing/Advanced_NLP_with_Python/04_Random_Forest_Model.ipynb <==
## NLP Basics: Building A Basic Random Forest Model On Top Of Vectorized Text

import pandas as pd
import string
from nltk.corpus import stopwords
import re

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import precision_score, recall_score

messages = pd.read_csv('data/spam.csv', encoding='latin-1')
messages.head()

messages = pd.read_csv('data/spam.csv', encoding='latin-1')
messages.head()

# Drop unused columns and label columns that will be used
messages = messages.drop(labels=["Unnamed: 2", "Unnamed: 3", "Unnamed: 4"], axis=1)
messages.columns = ['label', 'text']

messages.head()

messages.shape

### Create function to clean data

stopword = stopwords.words('english')

def clean_text(text):
    text = "".join([char.lower() for char in text if char not in string.punctuation])
    tokens = re.split('\W+', text)
    text = [word for word in tokens if word not in stopword]
    return text

### Apply TF-IDF Vectorizer

tfidf_vect = TfidfVectorizer(analyzer=clean_text)
X_tfidf_vect = tfidf_vect.fit_transform(messages['text'])

print(X_tfidf_vect.shape)
print(tfidf_vect.get_feature_names())

X_features = pd.DataFrame(X_tfidf_vect.toarray())

X_features.head()

### Explore RandomForestClassifier Attributes & Hyperparameters

#### Explre RandomForestClassifier on HoldOut set

X_train, X_test, y_train, y_test = train_test_split(X_features, messages['label'], test_size=0.2)

rfc = RandomForestClassifier()
rfc.fit(X_train, y_train)

y_pred = rfc.predict(X_test)

precision = precision_score(y_test, y_pred, pos_label='spam')
recall = recall_score(y_test, y_pred, pos_label='spam')
print(f"Precision: {round(precision, 3)} / Recall: {round(recall, 3)}")
