==> Natural_Language_Processing/Advanced_NLP_with_Python/05_Word2Vec_Part01.ipynb <==
# word2vec: How To Implement word2vec

### Explore Pre-trained Embeddings

Some other options:
- `glove-twitter-{25/50/100/200}`
- `glove-wiki-gigaword-{50/200/300}`
- `word2vec-google-news-300`
- `word2vec-ruscorpora-news-300`

# !pip install -U gensim

import gensim.downloader as api

wiki_emb = api.load('glove-wiki-gigaword-100')

# Explore the word vector for "king"
wiki_emb['king']

# Find the words most similar to king based on the trained word vectors
wiki_emb_emb_emb.most_similar("king")

### Train Our Own Model

import gensim
import numpy as np 
import pandas as pd
from sklearn.model_selection import train_test_split
pd.set_option('display.max_colwidth', 100)

messages = pd.read_csv('data/spam.csv', encoding='latin-1')
messages = messages.drop(labels=['Unnamed: 2', 'Unnamed: 3', 'Unnamed: 4'], axis=1)
messages.columns = ['label', 'text']
messages.head()

messages['text_clean'] = messages['text'].apply(lambda x: gensim.utils.simple_preprocess(x))
messages.head()

# Split data into train and test set
X_train, X_test, y_train, y_test = train_test_split(messages['text_clean'], messages['label'], test_size=0.2)

# Train Word2Vec Model
w2v_model = gensim.models.Word2Vec(X_train,
                      size=100,
                      window=5,
                      min_count=2)

w2v_model.wv['king']

w2v_model.wv.most_similar('king')
